requirements.txt
uvicorn
openai
python-dotenv
boto3
tenacity
numpy
groq
PyYAML
notebook
pydantic
fastapi
injector
requests